requirements.txt
asgiref==3.7.2
attrs==23.1.0
awsebcli==3.20.7
bcrypt==4.0.1
black==23.7.0
blessed==1.20.0
boto3==1.21.3
botocore==1.24.46
cached-property==1.5.2
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.3
cryptography==41.0.2
Django==4.2.3
django-cors-headers
django-crispy-forms==1.14.0
django-storages==1.12.3
django-tables2==2.4.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
jmespath==0.10.0
jsonschema==3.2.0
mypy-extensions==1.0.0
packaging==23.1
paramiko==3.2.0
pathspec==0.10.1
Pillow==10.0.0
platformdirs==3.8.1
psycopg2-binary==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.31.0
s3transfer==0.5.2
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
texttable==1.6.7
urllib3==1.26.16
wcwidth==0.1.9
websocket-client==0.59.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2